+++ react-materials/scaffolds/ice-scroll-screen-homepage/.eslintignore
@@ -1,0 +1,11 @@
+# 忽略目录
+build/
+tests/
+demo/
+
+# node 覆盖率文件
+coverage/
+
+# 忽略文件
+**/*-min.js
+**/*.min.js
