--- react-materials/scaffolds/ice-scroll-screen-homepage/src/components/BasicNotFound/index.js
@@ -1,3 +1,0 @@
-import BasicNotFound from './BasicNotFound';
-
-export default BasicNotFound;
