+++ react-materials/scaffolds/ice-scroll-screen-homepage/src/components/Logo/index.jsx
