--- react-materials/scaffolds/ice-scroll-screen-homepage/src/layouts/HeaderFooterLayout/index.js
@@ -1,3 +1,0 @@
-import HeaderFooterLayout from './HeaderFooterLayout';
-
-export default HeaderFooterLayout;
