+++ react-materials/scaffolds/ice-scroll-screen-homepage/src/menuConfig.js
@@ -2,13 +2,7 @@
 // headerMenuConfig：头部导航配置
 // asideMenuConfig：侧边导航配置
 
-const headerMenuConfig = [
-  {
-    name: '首页',
-    path: '/',
-  },
-];
-
+const headerMenuConfig = [];
 const asideMenuConfig = [];
 
 export { headerMenuConfig, asideMenuConfig };
