--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/components/AboutGeek/index.js
@@ -1,3 +1,0 @@
-import AboutGeek from './AboutGeek';
-
-export default AboutGeek;
