--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/components/Banner/index.js
@@ -1,3 +1,0 @@
-import Banner from './Banner';
-
-export default Banner;
