+++ react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/components/Banner/index.jsx
