--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/components/Events/index.js
@@ -1,3 +1,0 @@
-import Events from './Events';
-
-export default Events;
