--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/components/ScrollScreen/index.js
@@ -1,3 +1,0 @@
-import ScrollScreen from './ScrollScreen';
-
-export default ScrollScreen;
