--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/components/Sponsorship/index.js
@@ -1,3 +1,0 @@
-import Sponsorship from './Sponsorship';
-
-export default Sponsorship;
