--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/components/Summary/index.js
@@ -1,3 +1,0 @@
-import Summary from './Summary';
-
-export default Summary;
