--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/components/Workshop/index.js
@@ -1,3 +1,0 @@
-import Workshop from './Workshop';
-
-export default Workshop;
