--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/Home/index.js
@@ -1,3 +1,0 @@
-import Home from './Home';
-
-export default Home;
