--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/NotFound/NotFound.jsx
@@ -1,19 +1,0 @@
-import React, { Component } from 'react';
-import BasicNotFound from '../../components/BasicNotFound';
-
-export default class NotFound extends Component {
-  static displayName = 'NotFound';
-
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    return (
-      <div style={{ minHeight: '100vh' }}>
-        <BasicNotFound />
-      </div>
-    );
-  }
-}
