--- react-materials/scaffolds/ice-scroll-screen-homepage/src/pages/NotFound/index.js
@@ -1,3 +1,0 @@
-import NotFound from './NotFound';
-
-export default NotFound;
