+++ react-materials/scaffolds/ice-scroll-screen-homepage/src/routerConfig.js
@@ -2,20 +2,12 @@
 // 你可以调整 routerConfig 里的内容
 // 变量名 routerConfig 为 iceworks 检测关键字，请不要修改名称
 
-import HeaderFooterLayout from './layouts/HeaderFooterLayout';
 import Home from './pages/Home';
-import NotFound from './pages/NotFound';
 
 const routerConfig = [
   {
     path: '/',
-    layout: HeaderFooterLayout,
     component: Home,
-  },
-  {
-    path: '*',
-    layout: HeaderFooterLayout,
-    component: NotFound,
   },
 ];
